Rename file to nahmii-compiler and rename all references.
