+++ src/types/index.ts
@@ -2,4 +2,4 @@ export * from "./ThemeType"
 export * from "./input"
 export * from "./output"
 export * from "./compiler"
-export * from "./optimism-compiler"
+export * from "./nahmii-compiler"
